Normalize all line endings to LF in the repo
Change the CRLF strategie:
All (nearly all) files have LF in the repo, and CRLF in the working tree.
When we started, all files had been commited "as is".
Now it is more and more disturbing that Git shows "^M" in git diff
when a line was added or chenged.
To avoid this, store all files in the repo with LF.

To normalzie existing repos for a crate, run:

Step 1)
git fetch origin

Step 2)
git merge origin/master  -X renormalize